backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
python-multipart
google-genai
requests
fpdf2